modified mounting of mongo docker
Changed the mounting of the data/db/ directory in the docker to be in
the /var/datadir/ directory of the host instead of the /tmp/datadir/
directory. Should work on Linux with a running Docker machine etc;
though on WIndows 10 with a VirtualBox running Docker (this is the
development computer) the files need to be copied to the /var/ directory
everytime after a restart of the computer.